Change: Simplify internal implementation
