+++ source/TrainManager/SafetySystems/Plugin/ProxyPlugin.cs
@@ -1,5 +1,9 @@
-﻿using System.Diagnostics;
+﻿using System;
+using System.Diagnostics;
 using System.Runtime.InteropServices;
+using System.ServiceModel;
+using OpenBveApi.Hosts;
+using OpenBveApi.Interop;
 using OpenBveApi.Runtime;
 using SoundManager;
 using TrainManager.Trains;
@@ -22,13 +26,33 @@ namespace TrainManager.SafetySystems {
 		private GCHandle PanelHandle;
 		private GCHandle SoundHandle;
 
-		private readonly Win32ProxyPlugin pluginProxy = new Win32ProxyPlugin();
+		private readonly IAtsPluginProxy pipeProxy;
+		private readonly Process hostProcess;
+		private string lastError;
+		private bool externalCrashed;
 
 		// --- constructors ---
-		internal ProxyPlugin(string pluginFile, TrainBase train) {
+		internal ProxyPlugin(string pluginFile, TrainBase train)
+		{
+			externalCrashed = false;
 			base.PluginTitle = System.IO.Path.GetFileName(pluginFile);
 			//Load the plugin via the proxy callback
-			pluginProxy.setPluginFile(pluginFile, Process.GetCurrentProcess().Id);
+			hostProcess = new Process();
+			var handle = Process.GetCurrentProcess().MainWindowHandle;
+			try
+			{
+				hostProcess.StartInfo.FileName = @"Win32PluginProxy.exe";
+				hostProcess.Start();
+				HostInterface.Win32PluginHostReady.WaitOne();
+				pipeProxy = new DuplexChannelFactory<IAtsPluginProxy>(new InstanceContext(this), new NetNamedPipeBinding(), new EndpointAddress(HostInterface.Win32PluginHostEndpointAddress)).CreateChannel();
+				pipeProxy.SetPluginFile(pluginFile, Process.GetCurrentProcess().Id);
+				SetForegroundWindow(handle.ToInt32());
+			}
+			catch
+			{
+				//That didn't work
+				externalCrashed = true;
+			}
 			base.PluginValid = true;
 			base.PluginMessage = null;
 			base.Train = train;
@@ -48,12 +72,18 @@ namespace TrainManager.SafetySystems {
 			this.SoundHandle = new GCHandle();
 		}
 
-		
+		[DllImport("User32.dll")]
+		public static extern Int32 SetForegroundWindow(int hWnd);
 		
 		// --- functions ---
 		public override bool Load(VehicleSpecs specs, InitializationModes mode)
 		{
-			if (pluginProxy.load(specs, mode))
+			if (externalCrashed)
+			{
+				//Most likely the plugin proxy app failed to launch or something
+				return false;
+			}
+			if (pipeProxy.Load(specs, mode))
 			{
 				UpdatePower();
 				UpdateBrake();
@@ -69,10 +99,10 @@ namespace TrainManager.SafetySystems {
 			if (SoundHandle.IsAllocated) {
 				SoundHandle.Free();
 			}
-			pluginProxy.unload();
+			pipeProxy.Unload();
 		}
 		public override void BeginJump(InitializationModes mode) {
-			pluginProxy.beginJump(mode);
+			pipeProxy.BeginJump(mode);
 		}
 
 		public override void EndJump()
@@ -82,7 +112,7 @@ namespace TrainManager.SafetySystems {
 
 		protected override void Elapse(ref ElapseData data)
 		{
-			if (pluginProxy.externalCrashed)
+			if (externalCrashed)
 			{
 				//Yuck
 				for (int i = 0; i < Train.Cars[Train.DriverCar].Sounds.Plugin.Length; i++)
@@ -95,119 +125,120 @@ namespace TrainManager.SafetySystems {
 				Train.UnloadPlugin();
 				return;
 			}
-			if (pluginProxy.callback.lastError != string.Empty)
+			if (!string.IsNullOrEmpty(lastError))
 			{
 				
 				//TrainManagerBase.currentHost.A("ERROR: The proxy plugin " + PluginFile + " generated the following error:");
 				//Program.FileSystem.AppendToLogFile(pluginProxy.callback.lastError);
-				pluginProxy.callback.lastError = string.Empty;
+				lastError = string.Empty;
 			}
 
-			if (pluginProxy.callback.Unload == true)
+			try
 			{
-				return;
-			}
-			ElapseProxy e = new ElapseProxy(data, this.Panel, this.Sound);
-			ElapseProxy proxyData = pluginProxy.elapse(e);
-			this.Panel = proxyData.Panel;
-			this.Sound = proxyData.Sound;
-						for (int i = 0; i < this.Sound.Length; i++)
-			{
-				if (this.Sound[i] != this.LastSound[i])
+				ElapseProxy e = new ElapseProxy(data, this.Panel, this.Sound);
+				ElapseProxy proxyData = pipeProxy.Elapse(e);
+				this.Panel = proxyData.Panel;
+				this.Sound = proxyData.Sound;
+				for (int i = 0; i < this.Sound.Length; i++)
 				{
-					if (this.Sound[i] == SoundInstructions.Stop)
+					if (this.Sound[i] != this.LastSound[i])
 					{
-						if (i < base.Train.Cars[base.Train.DriverCar].Sounds.Plugin.Length)
+						if (this.Sound[i] == SoundInstructions.Stop)
 						{
-							Train.Cars[Train.DriverCar].Sounds.Plugin[i].Stop();
-						}
-					}
-					else if (this.Sound[i] > SoundInstructions.Stop & this.Sound[i] <= SoundInstructions.PlayLooping)
-					{
-						if (i < base.Train.Cars[base.Train.DriverCar].Sounds.Plugin.Length)
-						{
-							SoundBuffer buffer = base.Train.Cars[base.Train.DriverCar].Sounds.Plugin[i].Buffer;
-							if (buffer != null)
+							if (i < base.Train.Cars[base.Train.DriverCar].Sounds.Plugin.Length)
 							{
-								double volume = (double) (this.Sound[i] - SoundInstructions.Stop) / (double) (SoundInstructions.PlayLooping - SoundInstructions.Stop);
-								if (Train.Cars[Train.DriverCar].Sounds.Plugin[i].IsPlaying)
+								Train.Cars[Train.DriverCar].Sounds.Plugin[i].Stop();
+							}
+						}
+						else if (this.Sound[i] > SoundInstructions.Stop & this.Sound[i] <= SoundInstructions.PlayLooping)
+						{
+							if (i < base.Train.Cars[base.Train.DriverCar].Sounds.Plugin.Length)
+							{
+								SoundBuffer buffer = base.Train.Cars[base.Train.DriverCar].Sounds.Plugin[i].Buffer;
+								if (buffer != null)
 								{
-									Train.Cars[Train.DriverCar].Sounds.Plugin[i].Source.Volume = volume;
-								}
-								else
-								{
-									Train.Cars[Train.DriverCar].Sounds.Plugin[i].Play(1.0, volume, Train.Cars[Train.DriverCar], true);
+									double volume = (double) (this.Sound[i] - SoundInstructions.Stop) / (double) (SoundInstructions.PlayLooping - SoundInstructions.Stop);
+									if (Train.Cars[Train.DriverCar].Sounds.Plugin[i].IsPlaying)
+									{
+										Train.Cars[Train.DriverCar].Sounds.Plugin[i].Source.Volume = volume;
+									}
+									else
+									{
+										Train.Cars[Train.DriverCar].Sounds.Plugin[i].Play(1.0, volume, Train.Cars[Train.DriverCar], true);
+									}
 								}
 							}
 						}
-					}
-					else if (this.Sound[i] == SoundInstructions.PlayOnce)
-					{
-						if (i < base.Train.Cars[base.Train.DriverCar].Sounds.Plugin.Length)
+						else if (this.Sound[i] == SoundInstructions.PlayOnce)
 						{
-							SoundBuffer buffer = base.Train.Cars[base.Train.DriverCar].Sounds.Plugin[i].Buffer;
-							if (buffer != null)
+							if (i < base.Train.Cars[base.Train.DriverCar].Sounds.Plugin.Length)
 							{
-								Train.Cars[Train.DriverCar].Sounds.Plugin[i].Play(1.0, 1.0, Train.Cars[Train.DriverCar], false);
+								SoundBuffer buffer = base.Train.Cars[base.Train.DriverCar].Sounds.Plugin[i].Buffer;
+								if (buffer != null)
+								{
+									Train.Cars[Train.DriverCar].Sounds.Plugin[i].Play(1.0, 1.0, Train.Cars[Train.DriverCar], false);
+								}
 							}
+
+							this.Sound[i] = SoundInstructions.Continue;
+						}
+						else if (this.Sound[i] != SoundInstructions.Continue)
+						{
+							this.PluginValid = false;
 						}
 
-						this.Sound[i] = SoundInstructions.Continue;
+						this.LastSound[i] = this.Sound[i];
 					}
-					else if (this.Sound[i] != SoundInstructions.Continue)
+					else
 					{
-						this.PluginValid = false;
-					}
-
-					this.LastSound[i] = this.Sound[i];
-				}
-				else
-				{
-					if ((this.Sound[i] < SoundInstructions.Stop | this.Sound[i] > SoundInstructions.PlayLooping) && this.Sound[i] != SoundInstructions.PlayOnce & this.Sound[i] != SoundInstructions.Continue)
-					{
-						this.PluginValid = false;
+						if ((this.Sound[i] < SoundInstructions.Stop | this.Sound[i] > SoundInstructions.PlayLooping) && this.Sound[i] != SoundInstructions.PlayOnce & this.Sound[i] != SoundInstructions.Continue)
+						{
+							this.PluginValid = false;
+						}
 					}
 				}
+				data = proxyData.Data;
 			}
-
-			data = proxyData.Data;
-
-			data = proxyData.Data;
+			catch
+			{
+				lastError = externalCrashed.ToString();
+				externalCrashed = true;
+			}
 		}
 
 		protected override void SetReverser(int reverser) {
-			pluginProxy.setReverser(reverser);
+			pipeProxy.SetReverser(reverser);
 		}
 
 		protected override void SetPower(int powerNotch) {
-			pluginProxy.setPowerNotch(powerNotch);
+			pipeProxy.SetPowerNotch(powerNotch);
 		}
 
 		protected override void SetBrake(int brakeNotch) {
-			pluginProxy.setBrake(brakeNotch);
+			pipeProxy.SetBrake(brakeNotch);
 		}
 		public override void KeyDown(VirtualKeys key) {
-			pluginProxy.keyDown(key);
+			pipeProxy.KeyDown(key);
 		}
 		public override void KeyUp(VirtualKeys key) {
-			pluginProxy.keyUp(key);
+			pipeProxy.KeyUp(key);
 		}
 		public override void HornBlow(HornTypes type) {
-			pluginProxy.hornBlow(type);
+			pipeProxy.HornBlow(type);
 		}
 		public override void DoorChange(DoorStates oldState, DoorStates newState) {
-			pluginProxy.doorChange(oldState, newState);
+			pipeProxy.DoorChange(oldState, newState);
 		}
 
 		protected override void SetSignal(SignalData[] signal) {
 			if (base.LastAspects.Length == 0 || signal[0].Aspect != base.LastAspects[0])
 			{
-				pluginProxy.setSignal(signal[0].Aspect);
+				pipeProxy.SetSignal(signal[0].Aspect);
 			}
 		}
 
 		protected override void SetBeacon(BeaconData beacon) {
-			pluginProxy.setBeacon(beacon);
+			pipeProxy.SetBeacon(beacon);
 		}
 
 		protected override void PerformAI(AIData data)
